Split videos/show video player into two partials
Splits `videos/show`'s video player section into two partials since
there are so many checks for `signed_in`?
