+++ app/views/videos/_video_player_placeholder.html.erb
@@ -1,0 +1,9 @@
+<%= render "videos/access_callout", video: video %>
+
+<span class="divider">
+  <h4 class="text">Video</h4>
+</span>
+
+<div class="video-player">
+  <%= render "preview_clip_or_image", video: video %>
+</div>
